Runs a typical Seurat workflow on a Seurat object (up to dimensionality reduction and clustering).

Given a Seurat object, returns a new Seurat that has been normalized, had variable features identified, scaled, had principal components computed, hadclusters identified, and had tSNE and UMAP embeddings determined.

Usage

FindClustersCallback(
  seurat_obj,
  resolution_start = 0.8,
  reduction_percentage = 0.2,
  num_clusters_start = 20,
  dims = 1:10,
  algorithm = "louvain",
  assay = "RNA",
  cores = 1,
  verbose = TRUE
)

Arguments

seurat_obj

The Seurat object that will be analyzed.

resolution_start

The starting resolution to be used for the clustering algorithm (Louvain and Leiden algorithms).

reduction_percentage

The amount that the starting parameter will be reduced by after each iteration (between 0 and 1).

num_clusters_start

The starting number of clusters to be used for the clustering algorithm (K-means and Hierarchical clustering algorithms).

dims

The dimensions to use as input features (i.e. 1:10).

algorithm

The clustering algorithm to be used.

assay

The assay to generate knockoffs from.

cores

The number of cores to compute marker genes in parallel.

verbose

Whether or not to show all logging.

Value

Returns a Seurat object where the idents have been updated with the clusters determined via the callback algorithm. Latest clustering results will be stored in the object metadata under callback_clusters'. Note that 'callback_clusters' will be overwritten ever time FindClustersKC is run.